refactor: 리팩토링 - 오타 - print 함수 - 타이머 로직
