+++ lib/app_timer.dart
@@ -28,19 +28,6 @@ class AppTimer extends ChangeNotifier with WidgetsBindingObserver {
   void timeCount() {
     if (_timeState < _maxTimeState) {
       _timeState += 1;
-      notifyListeners();
-    }
-    if (_timeState == _maxTimeState) {
-      _timeState = 0;
-      notifyListeners();
-    }
-    if (_timeState % _maxTimeState == 0) {
-      _count += 1;
-      notifyListeners();
-    }
-
-    if (_timeState < _maxTimeState) {
-      _timeState += 1;
     } else {
       _timeState = 0;
       _count += 1;
@@ -48,7 +35,7 @@ class AppTimer extends ChangeNotifier with WidgetsBindingObserver {
     notifyListeners();
 
     if (_count == _maxCount) {
-      _timer?.cancel();
+      _stopTimer();
     }
   }
 
@@ -58,8 +45,7 @@ class AppTimer extends ChangeNotifier with WidgetsBindingObserver {
   }
 
   void resetTimer() {
-    _timer?.cancel();
-    _timer = null;
+    _stopTimer();
     _timeState = 0;
     _count = 0;
     notifyListeners();
